requirements.txt
streamlit
python-dotenv
pandas
numpy
scikit-learn
matplotlib
Pillow           
reportlab        
langgraph
langchain
langchain-openai
langchain-chroma
chromadb
tiktoken         
kaggle
